+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+***2024-09-28***
+
+- ***Changed***: Move "what's loaded" from `window.SYSTEM` to `window.loader`
+
+
 ***2024-09-23***
 
 - ***Fixed***: A non-secure website cannot load because the Cache API does not exist.